Conflicts:
	fuellogger_project/bin/R.txt
	fuellogger_project/bin/classes.dex
	fuellogger_project/bin/classes/no/appfortress/database/CarDBHandler.class
	fuellogger_project/bin/classes/no/appfortress/database/RefillDBHandler.class
	fuellogger_project/bin/classes/no/appfortress/fuellogger/DatabaseActivity.class
	fuellogger_project/bin/classes/no/appfortress/fuellogger/MainActivity.class
	fuellogger_project/bin/classes/no/appfortress/fuellogger/MyVehiclesFragment$1.class
	fuellogger_project/bin/classes/no/appfortress/fuellogger/MyVehiclesFragment.class
	fuellogger_project/bin/classes/no/appfortress/fuellogger/R$id.class
	fuellogger_project/bin/classes/no/appfortress/fuellogger/R$layout.class
	fuellogger_project/bin/classes/no/appfortress/fuellogger/RegisterVehicleFragment.class
	fuellogger_project/bin/classes/no/appfortress/fuellogger/VehiclesFragment.class
	fuellogger_project/bin/resources.ap_
	fuellogger_project/gen/no/appfortress/fuellogger/R.java
	fuellogger_project/src/no/appfortress/fuellogger/MyVehiclesFragment.java

+++ fuellogger_project/src/no/appfortress/database/RefillDBHandler.java
@@ -179,7 +179,7 @@ public class RefillDBHandler {
 				.getColumnIndex(CarFeedEntry.COLUMN_CAR_MODEL));
 		int year = cursor.getInt(cursor
 				.getColumnIndex(CarFeedEntry.COLUMN_YEAR));
-		long odometer = cursor.getLong(cursor
+		int odometer = cursor.getInt(cursor
 				.getColumnIndex(CarFeedEntry.COLUMN_ODOMETER));
 		float fuelTank = cursor.getFloat(cursor
 				.getColumnIndex(CarFeedEntry.COLUMN_FUELTANK));
